+++ src/test/java/com/library/runners/CukesRunner.java
@@ -18,7 +18,7 @@ import org.junit.runner.RunWith;
         glue = "com/library/step_defs",
         dryRun = false,
         publish = true,
-        tags = "@B28G33-89"
+        tags = "@wip"
 )
